requirements.txt
# FastAPI 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# 配置管理
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# LangChain 相关
langchain>=0.1.0
langchain-community>=0.0.20

# OpenAI 整合 SDK (用以支援所有 OpenAI 格式相容的供應商)
langchain-openai>=0.1.0

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# SSE 支持
sse-starlette>=1.6.0

# 开发工具
httpx>=0.25.0
